requirements.txt
astroid==2.3.3
autopep8==1.5
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cycler==0.10.0
Flask==1.1.1
Flask-Cors==3.0.8
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
gsheets==0.4.1
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.2
oauth2client==4.1.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
rsa==4.0
scipy==1.4.1
six==1.14.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
